example test case with markov module
