+++ lib/typing.py
@@ -1,0 +1,5 @@
+"""Typing annotations."""
+
+import disnake
+
+ApplicationCommandInteraction = disnake.ApplicationCommandInteraction
