+++ pyproject.toml
@@ -29,6 +29,7 @@ httpx = "^0.27.0"
 openai = "^1.14.3"
 aiofiles = "^23.2.1"
 defusedxml = "^0.7.1"
+pytest = "^8.2.2"
 
 [tool.poetry.group.dev.dependencies]
 black = "^23.3.0"
@@ -72,7 +73,6 @@ line-length = 120
 indent-width = 4
 target-version = "py311"
 
-
 [tool.ruff.lint]
 select = ["ALL"]
 ignore = ["E501", "ANN101", "COM812", "S311"]
@@ -85,3 +85,6 @@ quote-style = "double"
 indent-style = "space"
 skip-magic-trailing-comma = false
 line-ending = "auto"
+
+[tool.pytest.ini_options]
+pythonpath = [".", "lib", "lib/markovify"]
