+++ requirements.txt
@@ -10,7 +10,7 @@ attrs==23.2.0 ; python_version >= "3.11" and python_version < "4.0"
 backports-tarfile==1.0.0 ; python_version >= "3.11" and python_version < "3.12"
 certifi==2024.2.2 ; python_version >= "3.11" and python_version < "4"
 charset-normalizer==2.1.1 ; python_version >= "3.11" and python_version < "4"
-colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
+colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
 dateparser==1.2.0 ; python_version >= "3.11" and python_version < "4.0"
 defusedxml==0.7.1 ; python_version >= "3.11" and python_version < "4.0"
 disnake==2.8.1 ; python_version >= "3.11" and python_version < "4.0"
@@ -28,18 +28,21 @@ httpcore==1.0.5 ; python_version >= "3.11" and python_version < "4.0"
 httpx==0.27.0 ; python_version >= "3.11" and python_version < "4.0"
 huggingface-hub==0.22.2 ; python_version >= "3.11" and python_version < "4.0"
 idna==3.7 ; python_version >= "3.11" and python_version < "4"
+iniconfig==2.0.0 ; python_version >= "3.11" and python_version < "4.0"
 jaraco-context==5.3.0 ; python_version >= "3.11" and python_version < "4.0"
 more-itertools==10.2.0 ; python_version >= "3.11" and python_version < "4.0"
 multidict==6.0.5 ; python_version >= "3.11" and python_version < "4.0"
 openai==1.18.0 ; python_version >= "3.11" and python_version < "4.0"
 packaging==24.0 ; python_version >= "3.11" and python_version < "4.0"
 pillow==10.3.0 ; python_version >= "3.11" and python_version < "4.0"
+pluggy==1.5.0 ; python_version >= "3.11" and python_version < "4.0"
 prettytable==3.6.0 ; python_version >= "3.11" and python_version < "4.0"
 pydantic==2.7.0 ; python_version >= "3.11" and python_version < "4.0"
 pydantic-core==2.18.1 ; python_version >= "3.11" and python_version < "4.0"
 pyowm==3.3.0 ; python_version >= "3.11" and python_version < "4.0"
 pysocks==1.7.1 ; python_version >= "3.11" and python_version < "4.0"
 pyspellchecker==0.7.3 ; python_version >= "3.11" and python_version < "4.0"
+pytest==8.2.2 ; python_version >= "3.11" and python_version < "4.0"
 python-dateutil==2.9.0.post0 ; python_version >= "3.11" and python_version < "4.0"
 pytz==2024.1 ; python_version >= "3.11" and python_version < "4.0"
 pyyaml==6.0.1 ; python_version >= "3.11" and python_version < "4.0"
